[chore #163276689]Implement tests and refactor UserController file
