+++ .coveralls.yml
@@ -1,3 +1,2 @@
 service_name: travis-pro
 repo_token: NyWbMFejRIeIgTeiWVVXfD2qtVp50FHjU
-
