+++ app/v1/middleware/userValidator.js
@@ -1,4 +1,4 @@
-class signupValidator {
+class userValidator {
 	/**
 	 *@description - Checks the request parameters for creating new meetup are of the right formart
 	 *
@@ -13,10 +13,13 @@ class signupValidator {
 		request.check('firstname', 'firstname is required').notEmpty();
 		request.check('firstname', 'firstname must be a string').isString();
 		request.check('lastname', 'lastname is required').notEmpty();
-		request.check('email', 'email must be a string').isString();
+		request.check('lastname', 'lastname must be a string').isString();
+		request.check('email', 'email must be a valid email').isEmail();
 		request.check('email', 'email is required').notEmpty();
-		request.check('password', 'password must be a string').isString();
+		request.check('phoneNumber', 'phoneNumber must be valid').isNumeric();
+		request.check('phoneNumber', 'phoneNumber must be valid').trim().isLength({ min: 11, max: 11 });
 		request.check('password', 'password is required').notEmpty();
+		request.check('password', 'password must be longer than 5 characters').isLength({ min: 5 });
 		request.check('username', 'username must be a string').isString();
 		request.check('username', 'username is required').notEmpty();
 		const errors = request.validationErrors();
@@ -39,4 +42,4 @@ class signupValidator {
 		return next();
 	}
 }
-export default signupValidator;
+export default userValidator;
