+++ app/v1/routes/routes.js
@@ -1,12 +1,13 @@
 import express from 'express';
-import meetupValidator from '../middleware/meetupValidation';
+import meetupValidator from '../middleware/meetupValidator';
+import userValidator from '../middleware/userValidator';
 import UserController from '../dbControllers/UserController';
 import MeetupController from '../dbControllers/MeetupController';
 
 const router = express.Router();
 
 router.route('/auth/signup')
-	.post(UserController.signup);
+	.post(userValidator.signupValidator, UserController.signup);
 
 router.route('/auth/signin')
 	.post(UserController.signin);
