+++ package.json
@@ -7,7 +7,7 @@
     "start": "babel-node app/app.js",
     "serve": "nodemon app.js --exec babel-node --presets env,stage-2 app.js",
     "build": "babel app --out-dir build",
-	"pretest": "export NODE_ENV=test || SET \"NODE_ENV=test\" && npm run migration",
+    "pretest": "export NODE_ENV=test || SET \"NODE_ENV=test\" && npm run migration",
     "test": "export NODE_ENV=test || SET \"NODE_ENV=test\" && nyc mocha --exit --compilers js:babel-core/register --timeout 30000 ./app/test/**/*.js",
     "dev-start": "babel-watch app/app.js",
     "lint": "eslint app/app.js",
@@ -52,6 +52,7 @@
     "babel-watch": "^2.0.8",
     "body-parser": "^1.18.3",
     "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
     "chai-http": "^4.2.1",
     "dotenv": "^6.2.0",
     "express": "^4.16.4",
